berger98_plots.m : added comments
Added comments to code. An issue exists with modelling the Helium concentration,
so those plots as well as the total density plots are not identical to the ones
in the paper. This could be because of an error in the coefficients. No changes
to this code expected in the near future.

Signed-off-by: Lukas Bystricky <lb13f@my.fsu.edu>